Completed Code Test for Spin Tech
Requirements
We need 5 circles in a row. They should be in the following order. Red, Green, Blue, Orange and Yellow.
When they are hovered, text should appear under the icon that states what the color is and all other
Icons in the row should fade out to a 50% opacity.
Once the icon is clicked it should appear in a row below the original row and become a square. When it
is hovered, it should have the same effect as the circles. State the color above the icon and fade out the
other icons in the row. When a square is clicked, the square should become a circle and be placed first in
the circle row.
Page should be responsive, have a black background except when viewing on mobile, background
should be white, Open Sans font should be used, the site title should be your name and date, center
content and code should be commented. Use any libraries you are comfortable with, just list them on
the page after the circles and squares in a “Made with” section